#303 refactored sift_atomunit testing, fixed bugs
